Add functions to preserve view state+
The app initializes by checking if there is a saved state in local storage. If so, it loads the state. The ‘view’ value is taken from a data-set value stored in the header button html which is accessed by an event listener placed on the window object. The localStorage state is updated and the ‘state’ object in the app is updated to reflect those new values.